Add support for truncating fixed-format lines
Some legacy inputs depend on the fact that the Fortran standard states
that fixed-format input lines should be of fixed length, normally 72
characters.  Because the input processor would ignore characters past
this point, the space could be used as a sort of comment block.  This
commit winds a `last_column` variable through the parsers.  If it is
set to zero, the last column restriction is not enforced.